requirements.txt
# Generated requirements
agno>=0.5.0
loguru>=0.7.0
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
